requirements.txt
alembic==1.7.7
anyio==3.7.1     
asgiref==3.7.2   
click==8.1.7     
colorama==0.4.6  
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.78.0
flake8==6.0.0
greenlet==3.0.0
h11==0.14.0
httptools==0.6.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
openpyxl==3.1.2
psycopg2==2.9.7
pycodestyle==2.10.0
pydantic==1.10.13
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.19.1
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.17.6
watchgod==0.8.2
websockets==11.0.3